+++ app/models/artboard.server.ts
@@ -1,9 +1,12 @@
-import { type Artboard } from '@prisma/client'
+import { type Design, type Artboard } from '@prisma/client'
 import {
+	ArtboardSelectedDesignsSchema,
+	type ArtboardSelectedDesignsType,
 	type findArtboardArgsType,
 	type selectArgsType,
 	type whereArgsType,
 } from '#app/schema/artboard'
+import { type designTypeEnum } from '#app/schema/design'
 import { type PrismaTransactionType, prisma } from '#app/utils/db.server'
 
 // use prisma extension to .save() or .delete()
@@ -40,5 +43,38 @@ export const findArtboardTransactionPromise = ({
 }) => {
 	return prisma.artboard.findFirst({
 		where: { id },
+	})
+}
+
+// only use in transactions
+export const updateArtboardSelectedDesignPromise = ({
+	artboard,
+	designId,
+	type,
+	prisma,
+}: {
+	artboard: Pick<Artboard, 'id' | 'selectedDesigns'>
+	designId: Design['id']
+	type: designTypeEnum
+	prisma: PrismaTransactionType
+}) => {
+	// parse the selectedDesigns of the artboard
+	const selectedDesigns = ArtboardSelectedDesignsSchema.parse(
+		JSON.parse(artboard.selectedDesigns),
+	) as ArtboardSelectedDesignsType
+
+	// set the key for the selectedDesigns object to update
+	const designKey = (type + 'Id') as keyof ArtboardSelectedDesignsType
+
+	// build the updated selectedDesigns object
+	const updatedSelectedDesigns = ArtboardSelectedDesignsSchema.parse({
+		...selectedDesigns,
+		[designKey]: designId,
+	})
+
+	// update the selectedDesigns for the artboard
+	return prisma.artboard.update({
+		where: { id: artboard.id },
+		data: { selectedDesigns: JSON.stringify(updatedSelectedDesigns) },
 	})
 }
